streaming_server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
kafka-python==2.0.2
